+++ @navikt/core/react/src/date/index.ts
@@ -1,19 +1,17 @@
 export { type DateInputProps } from "./DateInput";
 export {
-  type DatePickerProps,
   default as DatePicker,
+  type DatePickerProps,
 } from "./datepicker/DatePicker";
 export { type DatePickerStandaloneProps } from "./datepicker/DatePickerStandalone";
-export { type MonthPickerStandaloneProps } from "./monthpicker/MonthPickerStandalone";
 export {
   useDatepicker,
   useMonthpicker,
   useRangeDatepicker,
-  type RangeValidationT,
   type DateValidationT,
   type MonthValidationT,
+  type RangeValidationT,
 } from "./hooks";
-export {
-  default as MonthPicker,
-  type MonthPickerProps,
-} from "./monthpicker/MonthPicker";
+export { default as MonthPicker } from "./monthpicker/MonthPicker";
+export { type MonthPickerStandaloneProps } from "./monthpicker/MonthPickerStandalone";
+export { type MonthPickerProps } from "./monthpicker/types";
