+++ @navikt/core/react/src/date/monthpicker/MonthPicker.tsx
@@ -1,98 +1,15 @@
 import cl from "clsx";
 import React, { forwardRef, useRef, useState } from "react";
 import { RootProvider } from "react-day-picker";
-import { DateInputProps, MonthPickerInput } from "../DateInput";
-import { DateContext, SharedMonthProvider } from "../context";
-import { getLocaleFromString, Matcher } from "../utils";
-import MonthCaption from "./MonthCaption";
-import MonthPickerStandalone, {
-  MonthPickerStandaloneType,
-} from "./MonthPickerStandalone";
-import MonthSelector from "./MonthSelector";
-import { useId } from "../../util";
 import { Popover } from "../../popover";
-
-export interface MonthPickerProps extends React.HTMLAttributes<HTMLDivElement> {
-  /**
-   * Element monthpicker anchors to. Use <MonthPicker.Input /> for built-in toggle,
-   * or make your own with the open/onClose props
-   */
-  children?: React.ReactNode;
-  /**
-   * Classname for datepicker in popover
-   */
-  className?: string;
-  /**
-   * Classname for wrapper
-   */
-  wrapperClassName?: string;
-  /**
-   * The earliest month to start navigation.
-   */
-  fromDate?: Date;
-  /**
-   * The latest day to end  navigation.
-   */
-  toDate?: Date;
-  /**
-   * Changes monthpicker locale
-   * @default "nb" (norsk bokmål)
-   */
-  locale?: "nb" | "nn" | "en";
-  /**
-   * Display dropdown for choosing year. Needs `fromDate` + `toDate` to work.
-   * @default false
-   */
-  dropdownCaption?: boolean;
-  /**
-   * Apply the disabled modifier to the matching months. Uses a subset of React Day Picker Matcher type. https://react-day-picker.js.org/api/types/Matcher
-   */
-  disabled?: Matcher[];
-  /**
-   * Controlled selected-month
-   */
-  selected?: Date;
-  /**
-   * Default selected month.
-   */
-  defaultSelected?: Date;
-  /**
-   * Open state for user-controlled state. Component controlled by default
-   */
-  open?: boolean;
-  /**
-   * onClose callback for user-controlled state
-   */
-  onClose?: () => void;
-  /**
-   * onOpenToggle callback for user-controlled-state. Only called if `<MonthPicker.Input />` is used
-   */
-  onOpenToggle?: () => void;
-  /**
-   * Callback for user-controlled state
-   */
-  onMonthSelect?: (month?: Date) => void;
-  /**
-   * Used to set visible year programmatically. Component controlled by default
-   */
-  year?: Date;
-  /**
-   * Event fired when the user navigates between years.
-   */
-  onYearChange?: (y?: Date) => void;
-  /**
-   * Avoid using if possible!
-   * Changes what CSS position property to use
-   * You want to use "fixed" if parent wrapper has position relative, but you want popover to escape
-   * @default "absolute"
-   */
-  strategy?: "absolute" | "fixed";
-  /**
-   * Bubbles Escape keydown-event up trough DOM-tree. This is set to false by default to prevent closing components like Modal on Escape
-   * @default false
-   */
-  bubbleEscape?: boolean;
-}
+import { useId } from "../../util";
+import { MonthPickerInput } from "../DateInput";
+import { DateContext, SharedMonthProvider } from "../context";
+import { getLocaleFromString } from "../utils";
+import MonthCaption from "./MonthCaption";
+import MonthPickerStandalone from "./MonthPickerStandalone";
+import MonthSelector from "./MonthSelector";
+import { MonthPickerProps } from "./types";
 
 interface MonthPickerComponent
   extends React.ForwardRefExoticComponent<MonthPickerProps> {
@@ -106,14 +23,12 @@ interface MonthPickerComponent
    * />
    * ```
    */
-  Standalone: MonthPickerStandaloneType;
+  Standalone: typeof MonthPickerStandalone;
   /**
    * Custom TextField for MonthPicker
    * @see 🏷️ {@link DateInputProps}
    */
-  Input: React.ForwardRefExoticComponent<
-    DateInputProps & React.RefAttributes<HTMLInputElement>
-  >;
+  Input: typeof MonthPickerInput;
 }
 
 /**
